pyproject.toml
[tool.poetry]
name = "url-checker"
version = "0.1.0"
description = "A compliance monitoring application for websites mentioning Admiral Markets"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = {extras = ["email"], version = "^2.6.0"}
python-multipart = "^0.0.9"
beautifulsoup4 = "^4.12.2"
lxml = "^5.1.0"
requests = "^2.32.0"
pandas = "^2.2.0"
pinecone-client = "^3.0.0"
sentence-transformers = "^2.6.0"
langchain = "^0.1.0"
openai = "^1.14.0"
spacy = "^3.7.0"
python-dotenv = "^1.0.0"
pytest = "^8.0.0"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
black = "^24.0.0"
isort = "^5.13.0"
flake8 = "^7.0.0"
mypy = "^1.7.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true 